Use size_t instead of int for nlevels and nessential

+++ include/gate.hpp
@@ -19,8 +19,8 @@ class Gate {
   protected:
     Mat V_re, V_im; ///< Real and imaginary parts of V_target, non-vectorized, essential levels only.
 
-    std::vector<int> nessential; ///< Number of essential levels per oscillator.
-    std::vector<int> nlevels; ///< Total number of levels per oscillator.
+    std::vector<size_t> nessential; ///< Number of essential levels per oscillator.
+    std::vector<size_t> nlevels; ///< Total number of levels per oscillator.
     int mpirank_petsc; ///< MPI rank in PETSc communicator.
     int mpisize_petsc; ///< MPI size in PETSc communicator.
     int mpirank_world; ///< MPI rank in world communicator.
@@ -54,7 +54,7 @@ class Gate {
      * @param lindbladtype_ Type of Lindblad operators for open system dynamics.
      * @param quietmode Flag to suppress output messages (default: false).
      */
-    Gate(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time_, const std::vector<double>& gate_rot_freq, LindbladType lindbladtype_, bool quietmode=false);
+    Gate(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time_, const std::vector<double>& gate_rot_freq, LindbladType lindbladtype_, bool quietmode=false);
 
     virtual ~Gate();
 
@@ -96,7 +96,7 @@ class Gate {
  */
 class XGate : public Gate {
   public:
-    XGate(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    XGate(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~XGate();
 };
 
@@ -111,7 +111,7 @@ class XGate : public Gate {
  */
 class YGate : public Gate {
   public:
-    YGate(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    YGate(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~YGate();
 };
 
@@ -126,7 +126,7 @@ class YGate : public Gate {
  */
 class ZGate : public Gate {
   public:
-    ZGate(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    ZGate(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~ZGate();
 };
 
@@ -141,7 +141,7 @@ class ZGate : public Gate {
  */
 class HadamardGate : public Gate {
   public:
-    HadamardGate(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    HadamardGate(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~HadamardGate();
 };
 
@@ -158,7 +158,7 @@ class HadamardGate : public Gate {
  */
 class CNOT : public Gate {
     public:
-    CNOT(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    CNOT(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~CNOT();
 };
 
@@ -175,7 +175,7 @@ class CNOT : public Gate {
  */
 class SWAP: public Gate {
     public:
-    SWAP(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    SWAP(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~SWAP();
 };
 
@@ -188,7 +188,7 @@ class SWAP: public Gate {
  */
 class SWAP_0Q: public Gate {
     public:
-    SWAP_0Q(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    SWAP_0Q(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~SWAP_0Q();
 };
 
@@ -200,7 +200,7 @@ class SWAP_0Q: public Gate {
  */
 class CQNOT: public Gate {
     public:
-    CQNOT(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    CQNOT(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~CQNOT();
 };
 
@@ -211,7 +211,7 @@ class CQNOT: public Gate {
  */
 class QFT: public Gate {
   public:
-    QFT(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
+    QFT(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, bool quietmode=false);
     ~QFT();
 };
 
@@ -233,7 +233,7 @@ class FromFile: public Gate {
      * @param filename Path to file containing gate matrix data.
      * @param quietmode Flag to suppress output messages (default: false).
      */
-    FromFile(const std::vector<int>& nlevels_, const std::vector<int>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, const std::string& filename, bool quietmode=false);
+    FromFile(const std::vector<size_t>& nlevels_, const std::vector<size_t>& nessential_, double time, const std::vector<double>& rotation_frequencies_, LindbladType lindbladtype_, const std::string& filename, bool quietmode=false);
     ~FromFile();
 };
 
@@ -253,4 +253,4 @@ class FromFile: public Gate {
  * @param quietmode Flag to suppress output messages.
  * @return Gate* Pointer to the initialized gate object.
  */
-Gate* initTargetGate(const std::vector<std::string>& target_str, const std::vector<int>& nlevels, const std::vector<int>& nessential, double total_time, LindbladType lindbladtype, const std::vector<double>& gate_rot_freq, bool quietmode);
+Gate* initTargetGate(const std::vector<std::string>& target_str, const std::vector<size_t>& nlevels, const std::vector<size_t>& nessential, double total_time, LindbladType lindbladtype, const std::vector<double>& gate_rot_freq, bool quietmode);
